pyproject.toml
[project]
name = "sales-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "python-dotenv>=1.1.0",
    "python-telegram-bot==20.8",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.1",
]
